backend/requirements.txt
asgiref==3.8.1
boto3==1.34.136
botocore==1.34.136
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.40.0
dj-database-url==2.2.0
Django==5.0.6
django-cleanup==8.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.0
django-resized==1.0.2
django-storages==1.14.3
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
jmespath==1.0.1
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
